Added structs to reduce function arguments
EarthModelBinned had some functions with entirely way too many variables.  In this update, 3 structs were introduced, 2 of which are used in passing arguments between functions.  The 3rd, TrajConstants, is used as the struct of one of the protected variables in EarthModelBinned, and has values that are updated whenever the detector position is set.  This required a restructuring of the way SetDetPos was defined.