Correct Error in EarthModelBinned (#35)
* EarthModel class addition

Also includes EarthBin struct

* Add EarthModel link definition

* Add EarthTables and defaults

This is where the binned Earth model tables should be kept for the EarthModel class to use.

* Add PREM default binned Earth model table

This one is what I used to test my code, but it would be better to shift this so that 0 depth is at sea level.

* Fixing compiler error

* Finished FillPath function

Finished modifications that would calculate the neutrino trajectory through latitude/longitude bins, as well as depth bins.

* Restructured PREM Earth bin table

The finish of the FillPath function also included a restructuring of the Earth bin table format to better match what I have previously been given.

* Fixed compilation errors

The compiler didn't like M_PI not being defined, which is in math.h.

* Reorganized Earth model classes

Now, PremModel and the other Earth model class (name changed to EarthModelBinned) share a base class

* Remove old EarthModel class

Removing the .cxx file for the old class

* Remove old EarthModel class

Removing .h file for the old class

* Update EarthModelBase.cxx

Add warning message for negative values of detector radius in SetDetPos

Co-authored-by: Joao Coelho <joaoabcoelho@gmail.com>

* Update PremModel.h

Remove commented out line

Co-authored-by: Joao Coelho <joaoabcoelho@gmail.com>

* Update EarthModelBase.cxx

Formatting preference change

Co-authored-by: Joao Coelho <joaoabcoelho@gmail.com>

* Remove model_defaults.hpp

This file should not exist in the repository: it needs to be created with the makefile (maybe as a part of prem_default.hpp, instead of separately).

* Update scheme for 3D model defaults

Use prem_default.hpp for 3D model defaults by adding extra lines to the makefile producing it.

* Update Makefile

Fixed mistakes and modified filename for the default model file for the EarthModelBinned class

* Delete prem_default.txt

Step 1 of modifying filename for default for EarthModelBinned

* Create earth_binned_default.txt

Step 2 in modifying filename for default for EarthModelBinned

* Apply suggestions from code review

update default file name in documentation

* Added structs to reduce function arguments

EarthModelBinned had some functions with entirely way too many variables.  In this update, 3 structs were introduced, 2 of which are used in passing arguments between functions.  The 3rd, TrajConstants, is used as the struct of one of the protected variables in EarthModelBinned, and has values that are updated whenever the detector position is set.  This required a restructuring of the way SetDetPos was defined.

* Modified PremModel.cxx

Changed some of the code/comments to be closer to the original, since I ended up re-adding SetDetPos to PremModel.

* Update PremModel.cxx

I just undid a change I had made due to not remembering why it was different from the original PremModel.cxx, because it needs to be that way to compile.

* Fixed errors in updates

After updating the longitude/latitude calculation functions to use less arguments, there was a problem with the longitude calculations.  This fixes that problem.

---------

Co-authored-by: Joao Coelho <joaoabcoelho@gmail.com>